requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11.15
django-tastypie==0.14.1
idna==2.7
pkg-resources==0.0.0
python-dateutil==2.7.3
python-decouple==3.1
python-mimeparse==1.6.0
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
